sample-corpus-tools/pyproject.toml
[project]
name = "sample-corpus-tools"
version = "0.1.0"
description = "Setup tooling for sample library analysis and FluCoMa CLI inventory"
readme = "README.md"
authors = [
    { name = "Sam Schorb", email = "sam_schorb@hotmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "soundfile>=0.13.1",
    "tqdm>=4.67.1",
    "typer>=0.20.0",
]

[project.scripts]
sct = "sample_corpus_tools.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
